more work on the config class
