+++ tests/test_cache.py
@@ -4,28 +4,9 @@ from datetime import datetime, timedelta
 
 import pytest
 
+from tests.common import BytesIOPath
 from vapor.cache_handler import Cache
 from vapor.data_structures import AntiCheatData, AntiCheatStatus, Game
-
-
-class BytesIOPath:
-	def __init__(self, bytes_io):
-		self.bytes_io = bytes_io
-
-	def read_text(self):
-		self.bytes_io.seek(0)
-		return self.bytes_io.read().decode()
-
-	def write_text(self, text):
-		self.bytes_io.seek(0)
-		self.bytes_io.truncate()
-		self.bytes_io.write(text.encode())
-
-	def __enter__(self):
-		return self
-
-	def __exit__(self, exc_type, exc_value, traceback):
-		self.bytes_io.close()
 
 
 @pytest.fixture
