+++ vapor/exceptions.py
@@ -9,11 +9,14 @@ class UnauthorizedError(Exception):
 class PrivateAccountError(Exception):
 	"""If an account is set to private."""
 
+
 class ConfigFileNotReadError(Exception):
 	"""If a config value is set without the config being read."""
 
+
 class ConfigWriteError(Exception):
 	"""If there was an error while writing the config."""
+
 
 class ConfigReadError(Exception):
 	"""If there was an error while reading the config."""
